Refactor serial port API
I need to develop a way to communicate with the Arduino that let's me send and receive large amounts of data.  I haven't figured out yet how to let the Arduino know that what I am sending is a command or data. By refactoring the code I hope to gain a better understanding of what the problem really is.