Create frequency-to-register function template in preparation for the MAX2871 driver.
To step through the entire band of the chip you need to set each frequency. The
first way to think about this is to send each requested frequency from the PC.
The second is to send a requested range of frequencies along with the number of
steps. The first method *must* be used when calibrating the unit so it would be
more efficient to send the frequencies as 3 bytes, or 24 bit binary, instead of
8 bytes, or a 64 bit, string.