Preliminary changes being made to support changing the mode of the mux pin.
The main reason is to be able to read the lock status from the MAX chip's register 6. But gaining read access to reg6 provides much more information than that. Call that icing on the cake. As of this commit the currentRegisterValue array is dead code but will be required for efficient communications when reg6 becomes readable.