Add try-except blocks to port_open(), write_serial_conf() and read_serial_conf().
I wanted these three functions to provide the following: 1) The program should attempt to open a serial port on startup.  2) If a serial port was previously opened the program should re-open the same port. 3) Upon changing the port or baudrate the serial port should re-open with the new settings. 4) The program should only open a serial port if it exists.