Remove reference clock from sweep() function parameters
Selecting a new reference clock in the hardware causes significant
delays during a sweep and should only be changed when necessary.
Controlling the ref_clock separately will allow all the control
codes in the full_sweep_dict to be arranged so that the ref1 codes
are all looped through first and then a single change to ref2 will
allow all the ref2 codes to be looped through second.