Fix the list of sweep frequencies:
Using a list of frequencies lets us use the meta data of the
list to provide the same boundaries for both the sweep and
the plot without worrying about synchronizing the sweep
and the plot.

This is a list of numpy.float's generated by numpy.arange().
But we want to include the stop-value as the last frequency in
the list which the arange() function doesn't do. Because the
arange function doesn't include it we can carelessly append
the sweep stop frequency to the final value in the list.