Refactor plot_amplitude_data to be more pythonic
Originally written when I first started learning Python so the source code looked like a c-programmer wrote it. The for-loop for processing and variables were changed to be self documenting. Instead of using range() to create an index for reading the list values it now reads the list values and requires no index(ing).