Move serial reads into the sweep loop
Serial reads were in their own thread but there's a problem where the serial requests arrive at the controller far too quickly for it to process the data and then do an ADC read.

It is hoped that putting the reads in with the requests will synchronize the communication cycle which can then be moved into it's own thread.