Control the focus event when filling in the Start (MHz), Stop (MHz), and Step Size (kHz) controls.
The focus events of the controls cause the sweep list to be updated over and over again.  Add a check for comparing the current value of the control(s) to the last entered value of the control(s) and just ignore the update if the two values are the same.