Generate a list of input frequencies that will be used by the sweep function
The fastest method I have found is to take the start/stop/step values and convert them to strings while padding the right hand side to three decimal places with 0's. Then the decimal is stripped and the string converted back to an integer.  These integers are then used to slice the RFin list with the frequencies to be swept.