Move all ref_clock_2 frequencies to be programmed after the ref_clock_1 frequencies.
The ref_clocks require a significant settling time when switching from one to the other.  If the board is programmed with sequential frequencies then there will be frequent switching from ref1 to ref2 which would have to be dealt with by large delays in the code.  This rewrite makes it so that all ref1 frequencies are programmed first along with collecting the ADC readings which is then followed by ref2.

This complicates the plotting because there is no information contained in the ADC value as to which frequency it is associated with.  That was dealt with by using numpy.argsort() to find the order of the swept_freq_list and reorder the indexes of  the list of amplitudes from the ADC.