Fix problem with amplitude peaks showing up at the wrong frequency
The Problem:  When sweeping it appeared that amplitudes associated with parts of the sweep_list were being displayed at the wrong frequency or frequencies.

Error Located:  The set_LO1() function only sends an updated control code to the Spectrum Analyzer when the step value for LO1 changes.  But the LO1 control code also needed to be updated every time the ref_clock was switched.

Solution:  In the function set_reference_clock() add the line self.last_LO1_code = 0 to force the set_LO1() function to send an update.