Add a keyboard.listener to allow early termination of the current sweep.
pynput keyboard.listener was added so that on_release(key) could set
a SWEEP flag when the ESC key was pressed. This allows the user to
terminate the current sweep without waiting for the entire thing to
run its course. A full sweep of 3 million data points could take
more than an hour.