Add a member function to create the combined full_control file.
The unit is calibrated by taking the amplitudes across the entire
bandwidth for each of the reference oscillators. With the amplitude
information a decision can be made whether to use the control codes
from the ref1 control file or the ref2 control file. The full
control file is created by selecting the ref1 or ref2 control file
based on which one generated the lower amplitude.