Fix improper selection of control codes during calibration.
The sweep function was dropping the amplitude value when a given set of control codes failed to get the Spectrum Analyzer to lock.  The fix was to add a MAX value so that it would get ignored during the calibration phase and would be a visual indicator when running a sweep.