Filter the amplitude calibration files for better control code selection
METHOD: When deciding which of the 4 control codes is best suited
for the calibrated control file the decision is made based on the
lowest voltage found in the associated amplitude files.

PROBLEM: When 2 or more amplitudes are in the noise-floor then
selecting the proper control code becomes random.

SOLUTION: The 4 amplitude calibration files must be filtered first.
A simple Kalman-like filter is used with constants and then the 4
filtered amplitude lists are used for selecting the control codes.