Change low pass filter function lp_filt()
The original filter was using a weighted mean which was causing
a phase delay in the amplitude data. This allowed a lot of spurs
to get through because the filtered amplitude data was no longer
in alignment with the RFin frequency list.

Solution:
Changed lp_filt() to use a Centered Moving Average that keeps the
filtered amplitude data properly aligned.

Result:
Far less spurs after calibration is complete.