Reduce 100% CPU usage when performing a frequency sweep
Problem:
When sweeping, the CPU goes to 100% and the cooling fans to go to max.

Reason:
A serial read is performed every time through the while-loop even if
there aren't any bytes in the serial buffer.

Solution:
Check for the presence of any bytes in the serial.in_waiting buffer.
Now, if there are any bytes waiting, we can perform a serial.read().