Bug the FMN values for 6000 MHz resulted in a small frequency error
The issue is that F was being truncated rather than rounded so the error
calculation was off by a very small amount. 32 ppm in this case.

Using truncation was biasing toward lower values of F and introducing a
quantization error. Rounding is needed to reach the exact frequency if
it's at all possible